--- packages/plaiceholder/src/defaults.ts
@@ -1,9 +1,0 @@
-import type { IGetPlaiceholderOptions } from "./plaiceholder";
-
-export const defaults: IGetPlaiceholderOptions = {
-  size: 4,
-  format: ["png"],
-  brightness: 1,
-  saturation: 1.2,
-  removeAlpha: true,
-};
