--- packages/plaiceholder/src/types.ts
@@ -1,1 +1,0 @@
-export const ACCEPTED_FILE_TYPES = ["jpeg", "png"] as const;
